requirements.txt
Flask
pyswisseph
geopy
pytz
gunicorn
timezonefinder
NumPy
Matplotlib
datetime
plotly